+++ examples/readme.py
@@ -1,55 +1,3 @@
-
-
-<p align="center">
-  <img src="https://depeche-py.github.io/depeche-db/assets/logo-bg.png" width="200" />
-</p>
-
-# Depeche DB
-
-A library for building event-based systems on top of PostgreSQL
-
-[![Tests](https://github.com/depeche-py/depeche-db/actions/workflows/tests.yml/badge.svg)](https://github.com/depeche-py/depeche-db/actions/workflows/tests.yml)
-[![pypi](https://img.shields.io/pypi/v/depeche-db.svg)](https://pypi.python.org/pypi/depeche-db)
-[![versions](https://img.shields.io/pypi/pyversions/depeche-db.svg)](https://github.com/depeche-py/depeche-db)
-[![Docs](https://img.shields.io/badge/docs-here-green.svg)](https://depeche-py.github.io/depeche-db/)
-[![license](https://img.shields.io/github/license/depeche-py/depeche-db.svg)](https://github.com/depeche-py/depeche-db/blob/main/LICENSE)
-
----
-
-**Documentation**: [https://depeche-py.github.io/depeche-db/](https://depeche-py.github.io/depeche-db/)
-
-**Source code**: [https://github.com/depeche-py/depeche-db](https://github.com/depeche-py/depeche-db)
-
----
-
-Depeche DB is modern Python library for building event-based systems
-
-Key features:
-
-* Message store with optimistic concurrency control & strong ordering guarantees
-* Subscriptions with "at least once" semantics
-* No database polling
-
-## Requirements
-
-Python 3.9+
-
-SQLAlchemy 1.4 or 2+
-
-PostgreSQL 12+
-
-
-## Installation
-
-```bash
-pip install depeche-db
-# OR
-poetry add depeche-db
-```
-
-## Example
-
-```python
 import pydantic, sqlalchemy, uuid, datetime as dt
 
 from depeche_db import (
@@ -111,12 +59,3 @@ aggregated_stream.projector.run()
 subscription.runner.run()
 # MyHandlers.handle_message prints:
 # SubscriptionMessage(partition=2, position=0, stored_message=StoredMessage(...))
-
-```
-
-
-## Contribute
-
-Contributions in the form of issues, questions, feedback and pull requests are
-welcome. Before investing a lot of time, let me know what you are up to so
-we can see if your contribution fits the vision of the project.