+++ tests/aggregated_stream/test_stream_projection.py
@@ -69,10 +69,9 @@ def test_stream_projector_cutoff(db_engine, store_factory, stream_factory, accou
     thread.join()
 
     # assert that the new events are not in the projection
-    with db_engine.connect() as conn:
-        assert account.events[-1].event_id not in [
-            msg.message_id for msg in subject.read(conn, partition=1)
-        ]
+    assert account.events[-1].event_id not in [
+        msg.message_id for msg in subject.read(partition=1)
+    ]
 
 
 def test_stream_projector_locking(db_engine, store_factory, stream_factory):
@@ -86,13 +85,12 @@ def test_stream_projector_locking(db_engine, store_factory, stream_factory):
 
 
 def assert_stream_projection(stream, db_engine, account, account2):
-    with db_engine.connect() as conn:
-        assert {msg.message_id for msg in stream.read(conn, partition=1)}.union(
-            {msg.message_id for msg in stream.read(conn, partition=2)}
-        ) == {event.event_id for event in account.events + account2.events}
-        assert [msg.message_id for msg in stream.read(conn, partition=1)] == [
-            event.event_id for event in account.events
-        ]
-        assert [msg.message_id for msg in stream.read(conn, partition=2)] == [
-            event.event_id for event in account2.events
-        ]
+    assert {msg.message_id for msg in stream.read(partition=1)}.union(
+        {msg.message_id for msg in stream.read(partition=2)}
+    ) == {event.event_id for event in account.events + account2.events}
+    assert [msg.message_id for msg in stream.read(partition=1)] == [
+        event.event_id for event in account.events
+    ]
+    assert [msg.message_id for msg in stream.read(partition=2)] == [
+        event.event_id for event in account2.events
+    ]